ts + restructuration
