+++ components/sections/footer/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export const Footer = () => {
+  return (
+    <div>
+    </div>
+  )
+}
