+++ components/sections/header/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Navbar } from '../../ui';
+
+export const Header = () => {
+  return (
+    <div>
+      <Navbar />
+    </div>
+  )
+}
