+++ components/sections/index.tsx
@@ -1,0 +1,3 @@
+export * from './footer';
+export * from './header';
+export * from './layout';
