+++ components/sections/layout/index.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Header, Footer } from "..";
+
+interface Layout {
+  // props you want to pass to the component other than the children
+}
+export default function Layout(props){
+  return (
+        <>
+            <Header />
+              {props.children}
+            <Footer />
+        </>
+  );
+}
