+++ package.json
@@ -36,10 +36,12 @@
     "web3modal": "^1.9.4"
   },
   "devDependencies": {
+    "@types/react": "^17.0.37",
     "autoprefixer": "^10.4.0",
     "eslint": "7.32.0",
     "eslint-config-next": "12.0.4",
     "postcss": "^8.4.3",
-    "tailwindcss": "^2.2.19"
+    "tailwindcss": "^2.2.19",
+    "typescript": "^4.5.4"
   }
 }
