--- src/components/Chains/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Chains";
