--- src/components/Layout.js
@@ -1,14 +1,0 @@
-import Navbar from "./Navbar";
-
-
-const Layout = ({children}) => {
-  return (
-        <div className="content">
-            <Navbar />
-              {children}
-        </div>
-  );
-}
-
-export default Layout;
-
